backend/requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Google Gemini AI (cloud-based image generation)
# Requires 1.10+ for gemini-2.5-flash-image model and part.as_image() method
google-genai>=1.10.0

# Pose Detection (lightweight, runs locally)
mediapipe>=0.10.9
opencv-python>=4.9.0

# Image processing
pillow>=10.2.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0
boto3>=1.34.0

# PDF Generation
reportlab>=4.0.0

# Database
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
aiosqlite>=0.19.0
